2/29/20 Intake, Shooting Close

+++ src/main/java/com/ironpanthers/frc2020/commands/FullShooterSequence.java
@@ -2,7 +2,6 @@ package com.ironpanthers.frc2020.commands;
 
 import com.ironpanthers.frc2020.commands.ShiftConveyor.Direction;
 import com.ironpanthers.frc2020.commands.arm.ArmInterpolation;
-import com.ironpanthers.frc2020.commands.arm.ArmToTarget;
 import com.ironpanthers.frc2020.commands.arm.ArmToTargetLL;
 
 /*----------------------------------------------------------------------------*/
@@ -14,7 +13,6 @@ import com.ironpanthers.frc2020.commands.arm.ArmToTargetLL;
 /*----------------------------------------------------------------------------*/
 
 import com.ironpanthers.frc2020.commands.shooter.ShooterInterpolation;
-import com.ironpanthers.frc2020.commands.shooter.StopShooter;
 import com.ironpanthers.frc2020.commands.vision.TurnToTarget;
 import com.ironpanthers.frc2020.subsystems.Arm;
 import com.ironpanthers.frc2020.subsystems.ConveyorBelt;
@@ -33,7 +31,7 @@ public class FullShooterSequence extends SequentialCommandGroup {
      * Creates a new FullShooterSequence.
      */
 
-    public FullShooterSequence(SteeringAdjuster steerer, Drive drive, Arm arm, int target, Shooter shooter,
+    public FullShooterSequence(SteeringAdjuster steerer, Drive drive, Arm arm, double target, Shooter shooter,
             int threshold, ConveyorBelt conveyor, LimelightWrapper lWrapper) {
         // Add your commands in the super() call, e.g.
         // super(new FooCommand(), new BarCommand());super();
